% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.r
\name{plot.accept_reject}
\alias{plot.accept_reject}
\title{Plot Accept-Reject}
\usage{
\method{plot}{accept_reject}(
  x,
  color_observed_density = "blue",
  color_true_density = "black",
  color_bar = "blue",
  color_observable_point = "red",
  color_real_point = "green",
  alpha = 0.3,
  ...
)
}
\arguments{
\item{x}{An object of class \verb{accept reject}}

\item{color_observed_density}{Observed density color (continuous case)}

\item{color_true_density}{True density color (continuous case)}

\item{color_bar}{Bar chart fill color (discrete case)}

\item{color_observable_point}{Color of generated points (discrete case)}

\item{color_real_point}{Color of real probability points (discrete case)}

\item{alpha}{Bar chart transparency (discrete case)}

\item{...}{Additional arguments}
}
\description{
Inspects the probability function (discrete case) or probability density
(continuous case) by comparing the theoretical case with the observed one.
}
\examples{

accept_reject(
   n = 2000L,
   f = dbinom,
   continuous = FALSE,
   args_f = list(size = 5, prob = 0.5),
   xlim = c(0, 10)
) |> plot()

accept_reject(
  n = 1000L,
  f = dnorm,
  continuous = TRUE,
  args_f = list(mean = 0, sd = 1),
  xlim = c(-4, 4)
) |> plot()

}
