% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.r
\name{print.accept_reject}
\alias{print.accept_reject}
\title{Print method for accept_reject objects}
\usage{
\method{print}{accept_reject}(x, n_min = 10L, ...)
}
\arguments{
\item{x}{An accept_reject object}

\item{n_min}{Minimum number of observations to print}

\item{...}{Additional arguments}
}
\description{
Print method for accept_reject objects
}
\examples{
set.seed(0) # setting a seed for reproducibility
accept_reject(
   n = 2000L,
   f = dbinom,
   continuous = FALSE,
   args_f = list(size = 5, prob = 0.5),
   xlim = c(0, 10)
) |> print()
}
