% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qqplot.R
\name{qqplot}
\alias{qqplot}
\title{QQ-Plot
QQ-Plot between observed quantiles and theoretical quantiles.}
\usage{
qqplot(x, ...)
}
\arguments{
\item{x}{Object of the class \code{accept_reject} returned by the function
\code{accept_reject()}.}

\item{...}{Additional arguments to be passed to methods.}
}
\value{
An object of classes \code{gg} and \code{ggplot} with the QQ-Plot of
theoretical quantiles versus observed quantiles.
}
\description{
QQ-Plot
QQ-Plot between observed quantiles and theoretical quantiles.
}
\details{
Generic method to plot the QQ-Plot between observed quantiles and theoretical
quantiles. The generic method will call the specific method
\code{qqplot.accept_reject()}, which operates on objects of class accept_reject
returned by the function \code{accept_reject()}.
}
\seealso{
\code{\link[=accept_reject]{accept_reject()}}, \code{\link[=print.accept_reject]{print.accept_reject()}}, \code{\link[=plot.accept_reject]{plot.accept_reject()}}
and \code{\link[=inspect]{inspect()}}.
}
